requirements.txt
google-genai
python-dotenv
pillow
rich
requests
markdownify